+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.9] - 2023-01-12
+
 ### Added
 - Load subtitle from player options
 - Added fps data for event 'fpsupdate'